+++ packages/ui5-test-writer/templates/freestyle/webapp/test/1.120.0/unit/AllTests.js
@@ -1,0 +1,5 @@
+sap.ui.define([
+	"<%= appIdWithSlash %>/test/unit/controller/<%= viewName %>.controller"
+], function () {
+	"use strict";
+});